requirements.txt
Pillow==8.2.0
streamlit==0.71.0
joblib==0.13.2
plotly==4.8.2
shap==0.37.0
pandas==1.2.4
#matplotlib==3.4.1
#numpy==1.20.2

